added lower_bound and upper_bound methods.
lower_bound returns the index of the element that is just "below" a
given coordinate.

upper_bouund returns the index of the element that is
just "above" a given coordinate.

So, given a coordinate, lower_bound and upper_bound give the opposite
corners of the cube that contains the coordinate.