feat: added Field::set_f overloads
Added overloads for Field::set_f to provide more options. Field::set_f
now accepts a callable that takes one argument, which will be an array
of the coordinates to evaluate the field at.

So now the caller doesn't have to ask for the coordinate from the
coordinate system with the index array if they don't need to use the
index values for anything else.

Also added support for a boost::optional or s std::optional return
type. If the callable passed to set_f returns an optional, then the
field value is only set if the returned optional is set. This allows the
caller to conditionally set field values, which was not possible before.