refactor: moved header include path
Headers are still installed to include/libField, but this directory is
not longer added to the include directories. So users need to include
libField/Field.hpp now instead of Field.hpp.

Also added a Field.hpp file that sits next to libField/ and displays a
warning about the new header location.